// client/src/api/magnetsServer/generated/models/Product.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Magnets e-Commerce API
 * API for managing products, orders, and user accounts for the Magnets e-Commerce platform.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Product
 */
export interface Product {
    /**
     * Unique identifier for the product
     * @type {string}
     * @memberof Product
     */
    id?: string;
    /**
     * Name of the product
     * @type {string}
     * @memberof Product
     */
    name?: string;
    /**
     * Description of the product
     * @type {string}
     * @memberof Product
     */
    description?: string;
    /**
     * Price of the product
     * @type {number}
     * @memberof Product
     */
    price?: number;
    /**
     * Availability of the product
     * @type {boolean}
     * @memberof Product
     */
    inStock?: boolean;
}

/**
 * Check if a given object implements the Product interface.
 */
export function instanceOfProduct(value: object): value is Product {
    return true;
}

export function ProductFromJSON(json: any): Product {
    return ProductFromJSONTyped(json, false);
}

export function ProductFromJSONTyped(json: any, ignoreDiscriminator: boolean): Product {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'price': json['price'] == null ? undefined : json['price'],
        'inStock': json['inStock'] == null ? undefined : json['inStock'],
    };
}

export function ProductToJSON(json: any): Product {
    return ProductToJSONTyped(json, false);
}

export function ProductToJSONTyped(value?: Product | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'description': value['description'],
        'price': value['price'],
        'inStock': value['inStock'],
    };
}